Move mocks to be shared and fix discord-event-handler tests
