+++ test/mocks/discord.js/voice-channel.ts
@@ -1,0 +1,3 @@
+export class VoiceChannel {
+  public join = jest.fn();
+}
